+++ fjs/dev/package_json/module.f.ts
@@ -4,7 +4,7 @@
  *
  * @module
  */
-import { parse as parseJsonText } from '../../media/json/module.f.ts'
+import { parse as parseJsonText } from '../../media/json/module.f.mjs'
 import { option, record, string } from '../../types/rtti/module.f.mjs'
 import { validate as rttiValidate } from '../../types/rtti/validate/module.f.mjs'
 import type { ValidationError } from '../../types/rtti/common/types.ts'
