+++ fjs/emergent_testing/proof.f.ts
@@ -13,7 +13,7 @@ import type { Reporter } from './types.ts'
 import { run as mockRun } from '../effects/mock/module.f.mjs'
 import type { All, Await, Test, TestContext } from '../effects/node/types.ts'
 import { shouldLoad } from '../dev/module.f.mjs'
-import { parse as parseJson } from '../media/json/module.f.ts'
+import { parse as parseJson } from '../media/json/module.f.mjs'
 import { array, number as rttiNumber, or, string as rttiString } from '../types/rtti/module.f.mjs'
 import { parse as rttiParse } from '../types/rtti/parse/module.f.mjs'
 import type { Ts } from '../types/rtti/ts/types.ts'
