+++ fjs/mcp/evo/proof.f.ts
@@ -8,7 +8,7 @@ import { initEvo, evo, type Evo } from '../../cas/evo/module.f.ts'
 import { evoToolRegistry } from './module.f.ts'
 import type { ToolEntry, ToolsCallResult } from '../../protocol/mcp/module.f.ts'
 import type { Operation } from '../../effects/types.ts'
-import { parse as parseJson } from '../../media/json/module.f.ts'
+import { parse as parseJson } from '../../media/json/module.f.mjs'
 import { array, string as rttiString } from '../../types/rtti/module.f.mjs'
 import { parse as rttiParse } from '../../types/rtti/parse/module.f.mjs'
 import { unwrap } from '../../types/result/module.f.mjs'
