+++ fjs/media/json/proof.f.ts
@@ -1,4 +1,4 @@
-import { parse, setProperty, stringify } from './module.f.ts'
+import { parse, setProperty, stringify } from './module.f.mjs'
 import { sort } from '../../types/object/module.f.mjs'
 import { identity } from '../../types/function/module.f.mjs'
 import { assertEq } from '../../asserts/module.f.mjs'
