+++ fjs/media/json/schema/proof.f.ts
@@ -1,7 +1,10 @@
+import { type Unknown as JsonValue } from '../types.ts'
+
 import { boolean, number, string, bigint, unknown, array, record, or, option } from '../../../types/rtti/module.f.mjs'
-import { stringify, type Unknown as JsonValue } from '../module.f.ts'
-import { toJsonSchema, type Unknown, unknown as schemaUnknown } from './module.f.ts'
+import { stringify } from '../module.f.mjs'
 import { assert, assertEq } from '../../../asserts/module.f.mjs'
+
+import { toJsonSchema, type Unknown, unknown as schemaUnknown } from './module.f.ts'
 
 const serialize = (v: Unknown) => stringify(e => e)(v as unknown as JsonValue)
