+++ fjs/media/module.f.ts
@@ -45,7 +45,7 @@ import type { Vec } from '../types/bit_vec/types.ts'
 import { fromVec } from '../text/utf8/module.f.mjs'
 import { detectVec } from './type/module.f.mjs'
 import type { DetectMeta } from './type/types.ts'
-import { parse } from './json/module.f.ts'
+import { parse } from './json/module.f.mjs'
 import { assert, assertNotNullish } from '../asserts/module.f.mjs'
 import type { Struct } from '../types/rtti/types.ts'
 import type { Ts, Unknown } from '../types/rtti/ts/types.ts'
