+++ fjs/media/revision/proof.f.ts
@@ -1,5 +1,7 @@
+import type { Object as JsonObject } from '../json/types.ts'
+
 import { assert, assertEq } from '../../asserts/module.f.mjs'
-import type { Object as JsonObject } from '../json/module.f.ts'
+
 import { dialect, mediaType, isHash, validate, decodeText, encodeText, type LockMap } from './module.f.ts'
 
 // Valid cbase32 hashes (round-tripped in fjs/basen/cbase32/proof.f.mjs): single
