+++ fjs/protocol/mcp/module.f.ts
@@ -13,20 +13,24 @@
  *
  * @module
  */
-import { boolean, string, option, array, record, or } from '../../types/rtti/module.f.mjs'
-import { unknown, type Unknown } from '../../media/json/module.f.ts'
+
+import type { Unknown } from '../../media/json/types.ts'
 import type { Ts } from '../../types/rtti/ts/types.ts'
-import { pure, step } from '../../effects/module.f.mjs'
 import type { Operation, Effect } from '../../effects/types.ts'
-import { read, write } from '../../effects/memory/module.f.mjs'
 import type { Key, MemOp } from '../../effects/memory/types.ts'
+
+import { boolean, string, option, array, record, or } from '../../types/rtti/module.f.mjs'
+import { unknown } from '../../media/json/rtti.f.mjs'
+import { pure, step } from '../../effects/module.f.mjs'
+import { read, write } from '../../effects/memory/module.f.mjs'
+import { validate } from '../../types/rtti/validate/module.f.mjs'
+
 import {
     decodeRequest,
     rpcError, invalidRequest, invalidParams, methodNotFound,
     type Response, type Id, type RpcError,
     jsonrpc,
 } from '../json_rpc/module.f.ts'
-import { validate } from '../../types/rtti/validate/module.f.mjs'
 import { toJsonSchema } from '../../media/json/schema/module.f.ts'
 import type { Type } from '../../types/rtti/types.ts'
