+++ fjs/protocol/mcp/proof.f.ts
@@ -1,12 +1,14 @@
+import type { Effect, Operation } from '../../effects/types.ts'
+import type { MemOperationMap } from '../../effects/mock/types.ts'
+import type { Key, MemOp } from '../../effects/memory/types.ts'
+import type { Unknown } from '../../media/json/types.ts'
+
 import { assert, assertEq } from '../../asserts/module.f.mjs'
 import { pure, step } from '../../effects/module.f.mjs'
 import { eff } from '../../effects/eff/module.f.mjs'
-import type { Effect, Operation } from '../../effects/types.ts'
 import { run } from '../../effects/mock/module.f.mjs'
-import type { MemOperationMap } from '../../effects/mock/types.ts'
 import { asBase, asNominal, create, read } from '../../effects/memory/module.f.mjs'
-import type { Key, MemOp } from '../../effects/memory/types.ts'
-import type { Unknown } from '../../media/json/module.f.ts'
+
 import {
     type ToolsListParams, type ToolsListResult, type ToolsCallParams, type ToolsCallResult,
     type McpHandlers, type McpConfig, type McpSessionState,
