+++ fjs/protocol/mcp/stdio/module.f.ts
@@ -29,12 +29,15 @@
  *
  * @module
  */
-import { pure, step } from '../../../effects/module.f.mjs'
+
 import type { Effect, Operation } from '../../../effects/types.ts'
-import { readLine, write } from '../../../effects/node/module.f.mjs'
 import type { IoResult, Read, Write } from '../../../effects/node/types.ts'
+import type { Unknown } from '../../../media/json/types.ts'
+
+import { pure, step } from '../../../effects/module.f.mjs'
+import { readLine, write } from '../../../effects/node/module.f.mjs'
 import { tryUtf8 } from '../../../text/module.f.mjs'
-import { parse, stringify, type Unknown } from '../../../media/json/module.f.ts'
+import { parse, stringify } from '../../../media/json/module.f.mjs'
 import { sort } from '../../../types/object/module.f.mjs'
 import { internalError, jsonrpc, parseError, type Response } from '../../json_rpc/module.f.ts'
 import { error, ok } from '../../../types/result/module.f.mjs'
