+++ fjs/text/ascii/proof.f.ts
@@ -1,5 +1,5 @@
 import { one, range } from './module.f.mjs'
-import { stringify as jsonStringify } from '../../media/json/module.f.ts'
+import { stringify as jsonStringify } from '../../media/json/module.f.mjs'
 import { sort } from '../../types/object/module.f.mjs'
 import { assertEq } from '../../asserts/module.f.mjs'
