+++ fjs/types/btree/set/proof.f.ts
@@ -1,7 +1,7 @@
 import { set as setSet } from './module.f.mjs'
 import type { TNode } from '../types/types.ts'
 import { cmp } from '../../string/module.f.mjs'
-import { stringify } from '../../../media/json/module.f.ts'
+import { stringify } from '../../../media/json/module.f.mjs'
 import { sort } from '../../object/module.f.mjs'
 import { assertEq } from '../../../asserts/module.f.mjs'
