+++ fjs/types/byte_set/proof.f.ts
@@ -1,6 +1,8 @@
+import type { Unknown } from '../../media/json/types.ts'
+
 import { has, empty, set, setRange, unset, universe, complement, toRangeMap } from './module.f.mjs'
 import { every, countdown, map, toArray } from '../list/module.f.mjs'
-import { stringify as jsonStringify, type Unknown } from '../../media/json/module.f.ts'
+import { stringify as jsonStringify } from '../../media/json/module.f.mjs'
 import { sort } from '../object/module.f.mjs'
 import { assert, assertEq } from '../../asserts/module.f.mjs'
