+++ fjs/types/list/proof.f.ts
@@ -1,6 +1,8 @@
 import type { List } from './types.ts'
+import type { Unknown } from '../../media/json/types.ts'
+
 import { length, concat, countdown, cycle, drop, dropWhile, entries, every, filter, find, flat, flatMap, map, next, reduce, reverse, scan, some, take, takeWhile, toArray, zip, first, filterMap, isEmpty, equal, tryFold } from './module.f.mjs'
-import { stringify, type Unknown } from '../../media/json/module.f.ts'
+import { stringify } from '../../media/json/module.f.mjs'
 import { sort } from '../object/module.f.mjs'
 import { addition, strictEqual, reduceToScan } from '../function/operator/module.f.mjs'
 import { assert, assertEq, assertNotNullish } from '../../asserts/module.f.mjs'
