+++ fjs/types/sorted_list/proof.f.ts
@@ -1,5 +1,7 @@
+import type { Unknown } from '../../media/json/types.ts'
+
 import { find, merge } from './module.f.mjs'
-import { stringify, type Unknown } from '../../media/json/module.f.ts'
+import { stringify } from '../../media/json/module.f.mjs'
 import { sort } from '../object/module.f.mjs'
 import { toArray, countdown, length } from '../list/module.f.mjs'
 import { flip } from '../function/module.f.mjs'
